factor out insert/remove functions
