+++ renderer/menu.js
@@ -1,3 +1,5 @@
+var hotController = require('../renderer/hot.js');
+
 const {remote} = require('electron');
 const {Menu, MenuItem} = remote;
 
@@ -6,76 +8,42 @@ var menu = new Menu();
 var rowAbove = new MenuItem({
   label: 'Insert row above',
   click: function() {
-    var range = hot.getSelectedRange();
-    var start = Math.min(range.from.row, range.to.row);
-    hot.alter('insert_row', start);
-    hot.deselectCell();
+    hotController.insertRowAbove();
   }
 });
 
 var rowBelow = new MenuItem({
   label: 'Insert row below',
   click: function() {
-    var range = hot.getSelectedRange();
-    var end = Math.max(range.from.row, range.to.row);
-    hot.alter('insert_row', (end + 1));
-    hot.deselectCell();
+    hotController.insertRowBelow();
   }
 });
 
 var columnLeft = new MenuItem({
   label: 'Insert column left',
   click: function() {
-    var range = hot.getSelectedRange();
-    var start = Math.min(range.from.col, range.to.col);
-    hot.alter('insert_col', start);
-    hot.deselectCell();
+    hotController.insertColumnLeft();
   }
 });
 
 var columnRight = new MenuItem({
   label: 'Insert column right',
   click: function() {
-    var range = hot.getSelectedRange();
-    var end = Math.max(range.from.col, range.to.col);
-    hot.alter('insert_col', (end + 1));
-    hot.deselectCell();
+    hotController.insertColumnRight();
   }
 });
 
 var removeRow = new MenuItem({
   label: 'Remove row(s)',
   click: function() {
-    var range = hot.getSelectedRange();
-
-    var start = Math.min(range.from.row, range.to.row);
-    var end   = Math.max(range.from.row, range.to.row);
-
-    for (var row = start; row <= end; row++) {
-      // rows are re-indexed after each remove
-      // so always remove 'start'
-      hot.alter('remove_row', start);
-    }
-
-    hot.deselectCell();
+    hotController.removeRows();
   }
 });
 
 var removeCol = new MenuItem({
   label: 'Remove column(s)',
   click: function() {
-    var range = hot.getSelectedRange();
-
-    var start = Math.min(range.from.col, range.to.col);
-    var end   = Math.max(range.from.col, range.to.col);
-
-    for (var col = start; col <= end; col++) {
-      // cols are re-indexed after each remove
-      // so always remove 'start'
-      hot.alter('remove_col', start);
-    }
-
-    hot.deselectCell();
+    hotController.removeColumns();
   }
 });
